refactor(structure): Organize components in 'auth' and 'pages' folders
