+++ pages/auth/Protected.jsx
@@ -1,0 +1,13 @@
+import React from "react";
+import { Navigate } from "react-router-dom";
+import { UserAuth } from "../../src/context/AuthContext";
+const Protected = ({ children }) => {
+  const { user } = UserAuth();
+
+  if (!user) {
+    return <Navigate to="/" />;
+  }
+  return children;
+};
+
+export default Protected;
