+++ pages/auth/Signin.jsx
@@ -1,10 +1,11 @@
 import { signInWithEmailAndPassword, AuthErrorCodes  } from 'firebase/auth';
 import React, { useEffect, useState } from 'react'
 // import { GoogleButton } from 'react-google-button';
-import { UserAuth } from '../../context/AuthContext';
+import { UserAuth } from '../../src/context/AuthContext';
 import { useNavigate, Link } from 'react-router-dom';
-import Logo from '../../assets/img/logo.png';
-import { auth } from '../../firebase';
+import Logo from '../../src/assets/img/logo.png'
+// import { auth } from '../../src/firebase';
+import { auth } from '../../src/firebase';
 
 
 function Signin() {
